Annual Maximum Sea Levels at Fremantle, Western Australia

This is a copy of the fremantle dataset from the ismev package. The fremantle data frame has 86 rows and 3 columns. The second column gives 86 annual maximum sea levels recorded at Fremantle, Western Australia, within the period 1897 to 1989. The first column gives the corresponding years. The third column gives annual mean values of the Southern Oscillation Index (SOI), which is a proxy for meteorological volatility.

Usage

fremantle

Format

This data frame contains the following:

Year: A numeric vector of years.

SeaLevel: A numeric vector of annual sea level maxima.

SOI: A numeric vector of annual mean values of the Southern Oscillation Index.

Source

Coles, S. G. (2001) An Introduction to Statistical Modelling of Extreme Values. London: Springer.

Examples

summary(fremantle)
#>       Year         SeaLevel          SOI          
#>  Min.   :1897   Min.   :1.190   Min.   :-1.78000  
#>  1st Qu.:1923   1st Qu.:1.460   1st Qu.:-0.44500  
#>  Median :1946   Median :1.520   Median : 0.06000  
#>  Mean   :1945   Mean   :1.538   Mean   :-0.03267  
#>  3rd Qu.:1968   3rd Qu.:1.620   3rd Qu.: 0.37250  
#>  Max.   :1989   Max.   :1.920   Max.   : 2.12000